Fix errors caused by renaming namespaces
