--- src/content/src/NetWebApiTemplate.Persistence/ISqlConnectionFactory.cs
@@ -1,9 +1,0 @@
-﻿using System.Data;
-
-namespace NetWebApiTemplate.Application.Shared.Interface
-{
-    public interface ISqlConnectionFactory
-    {
-        IDbConnection CreateConnection();
-    }
-}
